Fix compiler warning
I was getting "warning: cast from pointer to integer of different size". This is because the constant is actually coming from a #define, which means it's a literal number. You can't pass the address of a number, so make a local variable first and then pass the address of that.